+++ package.json
@@ -5,6 +5,7 @@
   "private": true,
   "scripts": {
     "tsc": "tsc",
+    "tsc:watch": "tsc --watch",
     "build": "webpack --config webpack.prod.js",
     "watch": "webpack --watch",
     "start": "webpack-dev-server --open --host --config webpack.dev.js",
@@ -55,11 +56,11 @@
     "express": "^4.17.1",
     "file-loader": "^6.1.0",
     "html-webpack-plugin": "^4.4.1",
-    "jest": "^25.5.4",
+    "jest": "^26.4.2",
     "jest-fetch-mock": "^3.0.3",
     "prettier": "^2.1.1",
     "style-loader": "^1.2.1",
-    "ts-jest": "^25.5.1",
+    "ts-jest": "^26.3.0",
     "typescript": "^4.0.2",
     "ts-loader": "^8.0.3",
     "webpack": "^4.44.1",
