+++ src/components/package/PackageData.ts
@@ -1,0 +1,12 @@
+import { IDependencyTreeConfig } from "../../../tools/npmdata/utils";
+
+export interface IPackageInfo {
+    name: string;
+    version: string;
+    description: string;
+    dependencies: number;
+    distinctDependencies: number;
+    directDependencies: number;
+    /** @deprecated */
+    tree: IDependencyTreeConfig;
+}
